% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping.R
\name{add_tiles}
\alias{add_tiles}
\title{Add tile data to the map}
\usage{
add_tiles(map, tiles = NULL)
}
\arguments{
\item{map}{leaflet::map}

\item{tiles}{Optional, currently select satellite for ARCGIS satellite imagery}
}
\value{
leaflet map
}
\description{
Add tile data to the map
}
\keyword{internal}
