% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_coords.R
\name{antimeridian_wrapping}
\alias{antimeridian_wrapping}
\title{Split polygons into two if crossing the dateline (antimeridian)}
\usage{
antimeridian_wrapping(
  polygon_data,
  crs = 4326,
  unique_inst = TRUE,
  to_sp = TRUE
)
}
\arguments{
\item{polygon_data}{SpatialPolygon or SpatialPolygonDataFrame object}

\item{crs}{CRS as a proj4string or EPSG code, default EPSG:4326 (datum=WGS84)}

\item{unique_inst}{Option to remove duplicate polygons (with the same coordinates), default TRUE}

\item{to_sp}{Option to convert POLYGON object into SpatialPolygon*, default TRUE}
}
\value{
SpatialPolygon or SpatialPolygonDataFrame split along the antimeridian
}
\description{
Split polygons into two if crossing the dateline (antimeridian)
}
