% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_paths.R
\name{clean_path}
\alias{clean_path}
\title{Returns a ~ expanded absolute path}
\usage{
clean_path(path, check_exists = FALSE)
}
\arguments{
\item{path}{Path to clean}

\item{check_exists}{Check if the path exists, error if it doesn't}
}
\value{
fs::path Expanded absolute path
}
\description{
Returns a ~ expanded absolute path
}
\keyword{internal}
