% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{create_config}
\alias{create_config}
\title{Create a user configuration file}
\usage{
create_config(tutorial_data = "fdmr_example_data", silent = FALSE)
}
\arguments{
\item{tutorial_data}{Path to tutorial data folder}

\item{silent}{Create the configuration file silently}
}
\description{
Create a user configuration file
}
