% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshing_internal.R
\name{create_mesh}
\alias{create_mesh}
\title{Create a mesh for ...}
\usage{
create_mesh(input_filepath, output_folder)
}
\arguments{
\item{input_filepath}{Path to input file}

\item{output_folder}{Output folder path}
}
\description{
Create a mesh for ...
}
