% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_sp}
\alias{fit_sp}
\title{Fit spatio-temporal data}
\usage{
fit_sp(
  formula,
  data,
  family,
  E,
  pred.link,
  reordering,
  int.strategy,
  mode,
  control.family = list(link = " "),
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Fit spatio-temporal data
}
