% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_shiny.R
\name{get_busy_spinner}
\alias{get_busy_spinner}
\title{Get a busy spinner for our Shiny apps.
If we have an internet connection we download the spinner
TODO - add this gif to the package}
\usage{
get_busy_spinner()
}
\value{
busyspinner
}
\description{
Get a busy spinner for our Shiny apps.
If we have an internet connection we download the spinner
TODO - add this gif to the package
}
\keyword{internal}
