% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_paths.R
\name{get_tutorial_datapath}
\alias{get_tutorial_datapath}
\title{Return the filepath for a tutorial data file.}
\usage{
get_tutorial_datapath(dataset, filename, saved = FALSE)
}
\arguments{
\item{dataset}{Name of dataset}

\item{filename}{Name of file}

\item{saved}{Specify the location where user unpacked the data.}
}
\value{
fs::path Full filepath
}
\description{
Return the filepath for a tutorial data file.
}
