% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_coords.R
\name{has_coords}
\alias{has_coords}
\title{Check if spatial data has coordinates set using sp::coordinates()
Returns TRUE if coordinates have been set}
\usage{
has_coords(spatial_data)
}
\arguments{
\item{spatial_data}{}
}
\value{
bool
}
\description{
Check if spatial data has coordinates set using sp::coordinates()
Returns TRUE if coordinates have been set
}
\keyword{internal}
