% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_priors.R
\name{interactive_priors}
\alias{interactive_priors}
\title{Interactively set and see the result of different priors}
\usage{
interactive_priors(spatial_data, measurement_data, time_variable, mesh)
}
\arguments{
\item{spatial_data}{Spatial data}

\item{measurement_data}{Measurement data}

\item{time_variable}{Time variable in measurement_data}

\item{mesh}{INLA mesh}
}
\value{
shiny::app
}
\description{
Interactively set and see the result of different priors
}
