% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_coords.R
\name{latlong_to_utm}
\alias{latlong_to_utm}
\title{Convert latitude and longitude to UTM coordinates}
\usage{
latlong_to_utm(lat, lon)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}
}
\value{
vector of UTM coordinates
}
\description{
Convert latitude and longitude to UTM coordinates
}
