% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_paths.R
\name{load_tutorial_data}
\alias{load_tutorial_data}
\title{Load data from the tutorial data store}
\usage{
load_tutorial_data(filename)
}
\arguments{
\item{filename}{Name of file}
}
\value{
loaded object
}
\description{
Load data from the tutorial data store
}
