% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_meshbuilder.R
\name{meshbuilder_shiny}
\alias{meshbuilder_shiny}
\title{Mesh building shiny app}
\usage{
meshbuilder_shiny(
  spatial_data,
  crs = NULL,
  max_edge = NULL,
  offset = NULL,
  cutoff = NULL
)
}
\arguments{
\item{spatial_data}{Spatial data}
}
\value{
shiny::app
}
\description{
Mesh building shiny app
}
\keyword{internal}
