% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modelviewer.R
\name{model_viewer}
\alias{model_viewer}
\title{Model viewing shiny app. Visualises INLA model output, plots the posterior distributions of model parameters and prediction statistics.}
\usage{
model_viewer(
  model_output,
  mesh,
  measurement_data,
  data_distribution = "Poisson"
)
}
\arguments{
\item{model_output}{INLA model output}

\item{mesh}{INLA mesh}

\item{measurement_data}{Measurement data}

\item{data_distribution}{Type of data, Poisson or Gaussian}
}
\value{
shiny::app
}
\description{
Model viewing shiny app. Visualises INLA model output, plots the posterior distributions of model parameters and prediction statistics.
}
