% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modelviewer.R
\name{model_viewer_shiny}
\alias{model_viewer_shiny}
\title{Parse inlabru model output}
\usage{
model_viewer_shiny(model_output, mesh, measurement_data, data_type)
}
\arguments{
\item{model_output}{INLA model output}

\item{mesh}{INLA mesh}

\item{measurement_data}{Measurement data}

\item{data_type}{Type of data, Poisson or Gaussian}
}
\value{
shiny::app
}
\description{
Parse inlabru model output
}
\keyword{internal}
