% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_strings.R
\name{numbers_only}
\alias{numbers_only}
\title{Removes any character that isn't a letter}
\usage{
numbers_only(data)
}
\arguments{
\item{data}{String data}
}
\value{
string data
}
\description{
Removes any character that isn't a letter
}
\keyword{internal}
