% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{parse_model_output_bru}
\alias{parse_model_output_bru}
\title{Parses inlabru::bru model output}
\usage{
parse_model_output_bru(model_output, measurement_data)
}
\arguments{
\item{model_output}{Output from running inlabru::bru}

\item{measurement_data}{Measurement data}
}
\value{
list
}
\description{
Parses inlabru::bru model output
}
\keyword{internal}
