% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_priors.R
\name{plot_ar1}
\alias{plot_ar1}
\title{Plot AR(1)}
\usage{
plot_ar1(data)
}
\arguments{
\item{data}{Parsed model output}

\item{to_plot}{Type of data to plot, "Range for f" etc}
}
\value{
ggplot2::ggplot
}
\description{
Plot AR(1)
}
\keyword{internal}
