% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot a boxplot using ggplot2}
\usage{
plot_boxplot(data, x, y, breaks, x_label, y_label)
}
\arguments{
\item{data}{Data to plot}

\item{x}{x-axis data}

\item{y}{y-axis data}

\item{breaks}{Break points}

\item{x_label}{x-axis label}

\item{y_label}{y-axis label}
}
\value{
ggplot
}
\description{
Plot a boxplot using ggplot2
}
