% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{plot_dic}
\alias{plot_dic}
\title{Plot Deviance Information Criterion (DIC) values}
\usage{
plot_dic(data)
}
\arguments{
\item{data}{List of parsed model outputs from "parse_model_output" (a list of lists)}
}
\value{
ggplot2::ggplot with DIC values shown as points
}
\description{
Plot Deviance Information Criterion (DIC) values
}
