% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mapper.R
\name{plot_interactive_map}
\alias{plot_interactive_map}
\title{Run the Shiny app for plotting raster and polygon data on a leaflet map.}
\usage{
plot_interactive_map(
  raster_data = NULL,
  polygon_data = NULL,
  date_format = NULL
)
}
\arguments{
\item{raster_data}{Raster data in the form or a RasterStack or RasterBrick}

\item{polygon_data}{Polygon data to plot on map}

\item{date_format}{A date format that will be passed in to lubridate::as_date}
}
\value{
shinyApp
}
\description{
Run the Shiny app for plotting raster and polygon data on a leaflet map.
}
