% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mapper.R
\name{plot_interative_map}
\alias{plot_interative_map}
\title{Run the interactive plotting Shiny app}
\usage{
plot_interative_map(raster_data = NULL, polygon_data = NULL)
}
\arguments{
\item{polygon_data}{}
}
\description{
Run the interactive plotting Shiny app
}
