% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_line_average}
\alias{plot_line_average}
\title{Create a line plot with a confidence interval.}
\usage{
plot_line_average(
  data,
  x,
  y1,
  y2,
  y3,
  breaks,
  x_label,
  y_label,
  y1_colour = "blue",
  y2_colour = "red",
  y3_colour = "red",
  x_lim = NULL,
  y_lim = NULL
)
}
\arguments{
\item{data}{Data to plot (all data types recognised by ggplot2::geom_line() like data.frame)}

\item{x}{x-axis data, must be a Date object}

\item{y1}{y1 line data - solid}

\item{y2}{y2 line data - dashed}

\item{y3}{y3 line data - dashed}

\item{breaks}{Breaks vector, must be a sequence of dates}

\item{x_label}{x-axis label}

\item{y_label}{y-axis label}

\item{y1_colour}{Colour for y1}

\item{y2_colour}{Colour for y2}

\item{y3_colour}{Colour for y3}

\item{x_lim}{x-axis limits as vector e.g. c(0, 0.1)}

\item{y_lim}{y-axis limits as vector e.g. c(0, 1.0)}
}
\value{
ggplot
}
\description{
Create a line plot with a confidence interval.
}
