% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping.R
\name{plot_map_leaflet}
\alias{plot_map_leaflet}
\title{Create a simple Leaflet map from data}
\usage{
plot_map_leaflet(
  polygon_data = NULL,
  raster_data = NULL,
  domain = NULL,
  markers = NULL,
  palette = "viridis",
  legend_title = NULL,
  add_scale_bar = FALSE,
  polygon_fill_colour = "#E4572E",
  polygon_line_colour = "grey",
  polygon_line_weight = 1,
  polygon_fill_opacity = 0.6,
  reverse = FALSE,
  wrapping = FALSE
)
}
\arguments{
\item{polygon_data}{Polygon data}

\item{raster_data}{Raster data}

\item{domain}{Domain data to be passed to leaflet::colorNumeric and leaflet::addLegend}

\item{markers}{Markers to display on map. A named list with latitude, longitude and label names must be given.}

\item{palette}{Palette to be used for colours, defaults to viridis}

\item{legend_title}{Title for legend}

\item{add_scale_bar}{Add scale bar if TRUE}

\item{polygon_fill_colour}{Polygon fill colour}

\item{polygon_line_colour}{Polygon surrounding line colour}

\item{polygon_line_weight}{Polygon surrounding line weight}

\item{polygon_fill_opacity}{Leaflet polygon fill opacity, float from 0 to 1.0, passed to fillOpacity of leaflet::addPolygons}

\item{reverse}{Reverse the colour palette if TRUE}

\item{wrapping}{Split polygons along the antimeridian (-180/180 boundary) if TRUE}
}
\value{
leaflet::leaflet
}
\description{
Create a simple Leaflet map from data
}
\keyword{internal}
