% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping.R
\name{plot_map_mapview}
\alias{plot_map_mapview}
\title{A simple map plotter using mapview. This is only intended for very quick viewing of data.}
\usage{
plot_map_mapview(polygon_data = NULL, raster_data = NULL)
}
\arguments{
\item{polygon_data}{Spatial data to plot}

\item{raster_data}{Raster data to plot}
}
\value{
mapview::mapview
}
\description{
A simple map plotter using mapview. This is only intended for very quick viewing of data.
}
\keyword{internal}
