% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping.R
\name{plot_map_polygon}
\alias{plot_map_polygon}
\title{Plot polygon data on a Leaflet map}
\usage{
plot_map_polygon(data, tiles = NULL, fill_colour = "red")
}
\arguments{
\item{data}{Polygon data}

\item{tiles}{Optional, currently select satellite for ARCGIS satellite imagery}

\item{fill_colour}{Fill colour of polygon}
}
\value{
leaflet map
}
\description{
Plot polygon data on a Leaflet map
}
