% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping.R
\name{plot_map_raster}
\alias{plot_map_raster}
\title{Plot raster data on a Leaflet Map}
\usage{
plot_map_raster(
  raster_data,
  tiles = NULL,
  polygon_data = NULL,
  polygon_fill_colour = "red"
)
}
\arguments{
\item{raster_data}{Raster data to plot}

\item{tiles}{Optional, currently select satellite for ARCGIS satellite imagery}
}
\value{
leaflet map
}
\description{
Plot raster data on a Leaflet Map
}
