% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mesh.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot a mesh on an interactive Leaflet map}
\usage{
plot_mesh(mesh)
}
\arguments{
\item{mesh}{Mesh data}
}
\value{
leaflet::leaflet
}
\description{
Plot a mesh on an interactive Leaflet map
}
