% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{plot_priors_boxplot}
\alias{plot_priors_boxplot}
\title{Create boxplots from priors run data}
\usage{
plot_priors_boxplot(data)
}
\arguments{
\item{data}{A list of parsed model output from "parse_model_output" (a list of lists)}
}
\value{
graphics::boxplot of fitted values
}
\description{
Create boxplots from priors run data
}
