% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{plot_priors_density}
\alias{plot_priors_density}
\title{Plot density function}
\usage{
plot_priors_density(data, measurement_data)
}
\arguments{
\item{data}{Parsed model outputs}

\item{measurement_data}{Measurement data}
}
\value{
ggplot2::ggplot
}
\description{
Plot density function
}
