% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mapper.R
\name{raster_mapping_app}
\alias{raster_mapping_app}
\title{Shiny app fo plotting raster data}
\usage{
raster_mapping_app(raster_data = NULL, polygon_data = NULL)
}
\arguments{
\item{polygon_data}{}
}
\description{
Shiny app fo plotting raster data
}
\keyword{internal}
