% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_packages.R
\name{require_package}
\alias{require_package}
\title{Checks if a package is installed. If it is it's loaded,
otherwise stop is called with a message telling the user
the package is required.}
\usage{
require_package(pkg_name)
}
\arguments{
\item{name}{Package name}
}
\description{
Checks if a package is installed. If it is it's loaded,
otherwise stop is called with a message telling the user
the package is required.
}
\keyword{internal}
