% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_packages.R
\name{require_packages}
\alias{require_packages}
\title{Checks if packages are installed.}
\usage{
require_packages(packages)
}
\arguments{
\item{packages}{Package name}
}
\description{
Checks if packages are installed.
}
\keyword{internal}
