% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dates.R
\name{to_dates}
\alias{to_dates}
\title{Convert vector of date strings to Date objects}
\usage{
to_dates(time_data, sort = FALSE, date_format = NULL)
}
\arguments{
\item{time_data}{Time data to sort}

\item{sort}{Sort if TRUE}

\item{date_format}{Date format to pass to lubridate::as_date}
}
\description{
Convert vector of date strings to Date objects
}
