% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_log.R
\name{write_log}
\alias{write_log}
\title{Write to a logfile}
\usage{
write_log(filepath, message)
}
\arguments{
\item{filepath}{Path to log}

\item{message}{Message to write}
}
\description{
Write to a logfile
}
\keyword{internal}
