% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_log.R
\name{write_parameters}
\alias{write_parameters}
\title{Write a parameter set to JSON}
\usage{
write_parameters(filepath, parameters)
}
\arguments{
\item{filepath}{Path to parameter file}

\item{message}{Parameters to write}
}
\description{
Write a parameter set to JSON
}
\keyword{internal}
