% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_log.R
\name{write_rds}
\alias{write_rds}
\title{A thin wrapper to write to an RDS file but also
checks to make sure we can write out to the filepath}
\usage{
write_rds(filepath, data, compress = TRUE)
}
\arguments{
\item{filepath}{Path to log}

\item{data}{Data to write}

\item{compress}{Compress the data}
}
\description{
A thin wrapper to write to an RDS file but also
checks to make sure we can write out to the filepath
}
\keyword{internal}
