% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{create_prediction_field}
\alias{create_prediction_field}
\title{Create a prediction field from the parsed model output and the mesh}
\usage{
create_prediction_field(
  mesh,
  plot_type = "predicted_mean_fields",
  data_dist = "poisson",
  var_a = NULL,
  var_b = NULL
)
}
\arguments{
\item{mesh}{INLA/fmesher mesh}

\item{plot_type}{Type of plot to create, "predicted_mean_fields" or "random_mean_fields"}

\item{data_dist}{Type of data, "poisson" or "gaussian"}

\item{var_a}{Data for variable a (numeric), required for "predicted_mean_fields" and "random_effect_fields"; normally, random effect field from "summary.random$f"}

\item{var_b}{Data for variable b (numeric), required for "predicted_mean_fields"; normally, fixed effect from "summary.fixed"}
}
\value{
data.frame with mesh node latlong coordinates (x, y) and the selected field values at the mesh nodes (z)
}
\description{
Create a prediction field from the parsed model output and the mesh
}
