% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{create_prediction_field}
\alias{create_prediction_field}
\title{Create a prediction field from the parsed model output and the mesh}
\usage{
create_prediction_field(var_a, var_b, mesh, crs = NULL)
}
\arguments{
\item{var_a}{Variable a data}

\item{var_b}{Variable b data}

\item{mesh}{INLA mesh}

\item{crs}{CRS as a proj4string}
}
\value{
data.frame
}
\description{
Create a prediction field from the parsed model output and the mesh
}
