% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_paths.R
\name{get_archive_cache_datapath}
\alias{get_archive_cache_datapath}
\title{Get path to downloaded archive cache folder}
\usage{
get_archive_cache_datapath()
}
\value{
fs::path
}
\description{
Get path to downloaded archive cache folder
}
\keyword{internal}
