% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_paths.R
\name{get_tutorial_cache_datapath}
\alias{get_tutorial_cache_datapath}
\title{Get path to tutorial data cache folder}
\usage{
get_tutorial_cache_datapath()
}
\value{
fs::path
}
\description{
Get path to tutorial data cache folder
}
\keyword{internal}
