% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_paths.R
\name{get_tutorial_datapath}
\alias{get_tutorial_datapath}
\title{Return the filepath for a tutorial data file.}
\usage{
get_tutorial_datapath(filename)
}
\arguments{
\item{filename}{Name of file}
}
\value{
fs::path Full filepath
}
\description{
Return the filepath for a tutorial data file.
}
