% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_meshbuilder.R
\name{mesh_builder}
\alias{mesh_builder}
\title{Mesh building shiny app. Creates and visualises a mesh from some spatial data.}
\usage{
mesh_builder(
  spatial_data,
  crs = NULL,
  max_edge = NULL,
  offset = NULL,
  cutoff = NULL
)
}
\arguments{
\item{spatial_data}{Spatial data}
}
\value{
shiny::app
}
\description{
Mesh building shiny app. Creates and visualises a mesh from some spatial data.
}
