% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_meshbuilder.R
\name{mesh_builder}
\alias{mesh_builder}
\title{Mesh building shiny app. Creates and visualises a mesh from some spatial data.}
\usage{
mesh_builder(
  spatial_data,
  obs_data = NULL,
  crs = NULL,
  max_edge = NULL,
  offset = NULL,
  cutoff = NULL,
  y_coord = "LAT",
  x_coord = "LONG"
)
}
\arguments{
\item{spatial_data}{Spatial data (must be a data.frame, SpatialPointsDataFrame or SpatialPolygonsDataFrame)}

\item{obs_data}{Measurement data}

\item{crs}{CRS as a proj4string}

\item{max_edge}{The largest allowed triangle edge length. One or two values, passed to fmesher::fm_mesh_2d_inla}

\item{offset}{Specifies the size of the inner and outer extensions around data locations, passed to fmesher::fm_mesh_2d_inla}

\item{cutoff}{The minimum allowed distance between points, passed to fmesher::fm_mesh_2d_inla}

\item{y_coord}{Name of the latitude column in the spatial data}

\item{x_coord}{Name of the longitude column in the spatial data}
}
\value{
shiny::app with meshbuilder functionality
}
\description{
Mesh building shiny app. Creates and visualises a mesh from some spatial data.
}
