% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshing_checks.R
\name{mesh_checker}
\alias{mesh_checker}
\title{Check a mesh for errors:
\enumerate{
\item A greater number of mesh nodes than observations
\item A greater number of mesh triangles than observations
\item Isolated triangles in mesh
}}
\usage{
mesh_checker(mesh, observations)
}
\arguments{
\item{mesh}{INLA mesh}

\item{observations}{Observations data to be used woith model}
}
\value{
list: Named list containing errors
}
\description{
Check a mesh for errors:
\enumerate{
\item A greater number of mesh nodes than observations
\item A greater number of mesh triangles than observations
\item Isolated triangles in mesh
}
}
