% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_meshbuilder.R
\name{meshbuilder_shiny}
\alias{meshbuilder_shiny}
\title{Mesh building shiny app}
\usage{
meshbuilder_shiny(
  spatial_data,
  obs_data = NULL,
  crs = NULL,
  max_edge = NULL,
  offset = NULL,
  cutoff = NULL
)
}
\arguments{
\item{spatial_data}{Spatial data}

\item{crs}{CRS as a proj4string}

\item{max_edge}{The largest allowed triangle edge length. One or two values, passed to inla.mesh.2d}

\item{offset}{Specifies the size of the inner and outer extensions around data locations, passed to inla.mesh.2d}

\item{cutoff}{The minimum allowed distance between points, passed to inla.mesh.2d}

\item{data}{Observations data, for use with the check_mesh functionality}
}
\value{
shiny::app
}
\description{
Mesh building shiny app
}
\keyword{internal}
