% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_meshbuilder.R
\name{meshbuilder_shiny}
\alias{meshbuilder_shiny}
\title{Mesh building shiny app}
\usage{
meshbuilder_shiny(location_data)
}
\arguments{
\item{location_data}{SpatialPolygon(ish) data}
}
\value{
shiny::app
}
\description{
Mesh building shiny app
}
\keyword{internal}
