% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_meshbuilder.R
\name{meshbuilder_shiny}
\alias{meshbuilder_shiny}
\title{Mesh building shiny app}
\usage{
meshbuilder_shiny(
  spatial_data,
  obs_data = NULL,
  crs = NULL,
  offset = NULL,
  max_edge = NULL,
  cutoff = NULL,
  plot_poly = FALSE,
  y_coord = "LAT",
  x_coord = "LONG"
)
}
\arguments{
\item{spatial_data}{Spatial data}

\item{obs_data}{Measurement data}

\item{crs}{CRS as a proj4string}

\item{offset}{Specifies the size of the inner and outer extensions around data locations, passed to fmesher::fm_mesh_2d_inla}

\item{max_edge}{The largest allowed triangle edge length. One or two values, passed to fmesher::fm_mesh_2d_inla}

\item{cutoff}{The minimum allowed distance between points, passed to fmesher::fm_mesh_2d_inla}

\item{y_coord}{Name of the latitude column in the spatial data}

\item{x_coord}{Name of the longitude column in the spatial data}
}
\value{
shiny::app
}
\description{
Mesh building shiny app
}
\keyword{internal}
