% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modelbuilder.R
\name{model_builder}
\alias{model_builder}
\title{Interactively set and see the result of different priors}
\usage{
model_builder(
  spatial_data,
  measurement_data,
  time_variable,
  mesh,
  data_distribution = "Poisson",
  log_folder = NULL
)
}
\arguments{
\item{spatial_data}{Spatial data}

\item{measurement_data}{Measurement data (data.frame, SpatialPointsDataFrame, SpatialPolygonsDataFrame with a named time column and at least 1 named feature column with measurements)}

\item{time_variable}{Time variable name in measurement_data}

\item{mesh}{INLA/fmesher mesh}

\item{log_folder}{Folder to write logs to (character)}
}
\value{
shiny::app with model builder functionality
}
\description{
Interactively set and see the result of different priors
}
