% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_priors.R
\name{model_builder_shiny}
\alias{model_builder_shiny}
\title{Interactively set and see the result of different priors}
\usage{
model_builder_shiny(
  spatial_data,
  measurement_data,
  time_variable,
  mesh,
  data_distribution = "Poisson",
  log_folder = NULL
)
}
\arguments{
\item{spatial_data}{Spatial data}

\item{measurement_data}{Measurement data}

\item{time_variable}{Time variable in measurement_data}

\item{mesh}{INLA mesh}

\item{data_distribution}{Data distribution, Poisson or Gaussian}

\item{log_folder}{Folder to write out logs}
}
\value{
shiny::app
}
\description{
Interactively set and see the result of different priors
}
\keyword{internal}
