% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modelviewer.R
\name{model_viewer}
\alias{model_viewer}
\title{Mesh building shiny app. Creates and visualises a mesh from some spatial data.}
\usage{
model_viewer(model_output, mesh, measurement_data, data_type = "Poisson")
}
\arguments{
\item{model_output}{INLA model output}

\item{mesh}{INLA mesh}

\item{measurement_data}{Measurement data}

\item{data_type}{Type of data, Poisson or Gaussian}
}
\value{
shiny::app
}
\description{
Mesh building shiny app. Creates and visualises a mesh from some spatial data.
}
