% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modelviewer.R
\name{model_viewer}
\alias{model_viewer}
\title{Mesh building shiny app. Creates and visualises a mesh from some spatial data.}
\usage{
model_viewer(
  model_output,
  mesh,
  measurement_data,
  data_distribution = "Poisson"
)
}
\arguments{
\item{model_output}{INLA/inlabru model output}

\item{mesh}{INLA/fmesher mesh used for the run}

\item{measurement_data}{Measurement data used for the run (can be data.frame, SpatialPoints/SpatialPointsDataFrame, or simple feature collection)}

\item{data_distribution}{Type of data, Poisson or Gaussian (character)}
}
\value{
shiny::app with model viewer functionality
}
\description{
Mesh building shiny app. Creates and visualises a mesh from some spatial data.
}
