% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modelviewer.R
\name{model_viewer_shiny}
\alias{model_viewer_shiny}
\title{Parse inlabru model output}
\usage{
model_viewer_shiny(model_output, mesh, measurement_data, data_distribution)
}
\arguments{
\item{model_output}{INLA/inlabru model output}

\item{mesh}{INLA/fmesher mesh}

\item{measurement_data}{Measurement data (can be data.frame, SpatialPoints/SpatialPointsDataFrame, or simple feature collection)}

\item{data_distribution}{Type of data, Poisson or Gaussian}
}
\value{
shiny::app with model viewer functionality
}
\description{
Parse inlabru model output
}
\keyword{internal}
