% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{parse_model_output}
\alias{parse_model_output}
\title{Parse model output}
\usage{
parse_model_output(model_output, measurement_data, model_type = "inlabru")
}
\arguments{
\item{model_output}{Data returned by model}

\item{measurement_data}{Measurement data}

\item{model_type}{Type of model, we currently support inlabru}
}
\value{
list
}
\description{
Parse model output
}
