% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{parse_model_output_bru}
\alias{parse_model_output_bru}
\title{Parses inlabru::bru model output to create a list of model parameters}
\usage{
parse_model_output_bru(model_output, measurement_data)
}
\arguments{
\item{model_output}{Output from running inlabru::bru}

\item{measurement_data}{Measurement data (can be data.frame, SpatialPoints/SpatialPointsDataFrame, or simple feature collection)}
}
\value{
Creates a list of model outputs for the first timestep, including fitted mean and sd, mean of fixed effects, mean and sd of random effects, posterior hyperparameters, and DIC
}
\description{
Parses inlabru::bru model output to create a list of model parameters
}
\keyword{internal}
