% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_barchart}
\alias{plot_barchart}
\title{Plot a bar chart using ggplot2}
\usage{
plot_barchart(
  data,
  x,
  y,
  breaks,
  x_label,
  y_label,
  fill = "pink",
  colour = "blue"
)
}
\arguments{
\item{data}{Data to plot}

\item{x}{x-axis data}

\item{y}{y-axis data}

\item{breaks}{Break points}

\item{x_label}{x-axis label}

\item{y_label}{y-axis label}

\item{fill}{Fill colour}

\item{colour}{Line colour}
}
\value{
ggplot
}
\description{
Plot a bar chart using ggplot2
}
