% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_priors.R
\name{plot_dic}
\alias{plot_dic}
\title{Plot Deviance Information Criterion (DIC) values}
\usage{
plot_dic(data)
}
\arguments{
\item{data}{}
}
\value{
ggplot2::ggplot
}
\description{
Plot Deviance Information Criterion (DIC) values
}
\keyword{internal}
