% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{plot_line_comparison}
\alias{plot_line_comparison}
\title{Plot line comparison for stdev etc}
\usage{
plot_line_comparison(data, to_plot, title)
}
\arguments{
\item{data}{Parsed model output}

\item{to_plot}{Type of data to plot, "Range for f" etc}
}
\value{
ggplot2::ggplot
}
\description{
Plot line comparison for stdev etc
}
\keyword{internal}
