% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping.R
\name{plot_map}
\alias{plot_map}
\title{Create a simple Leaflet map from data}
\usage{
plot_map(
  data,
  domain,
  palette = "YlOrRd",
  colour = "grey",
  legend_values = NULL,
  legend_title = NULL,
  add_scale_bar = FALSE,
  polygon_fill_opacity = 0.75,
  fill_colour_weight = 1
)
}
\arguments{
\item{data}{Data to plot}

\item{domain}{Domain for map}

\item{palette}{Palette, for example YlOrRd or Reds}

\item{legend_values}{Values for legend}

\item{legend_title}{Title for legend}

\item{add_scale_bar}{Add scale bar if TRUE}

\item{polygon_fill_opacity}{Leaflet polygon fill opacity, float from 0 to 1.0, passed to fillOpacity of leaflet::addPolygons}

\item{fill_colour_weight}{Polygon colour weight, float from 0 to 1.0, Passed to the weight argument of addPolygons}
}
\value{
map
}
\description{
Create a simple Leaflet map from data
}
