% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping.R
\name{plot_map}
\alias{plot_map}
\title{Create a simple Leaflet map from data}
\usage{
plot_map(
  polygon_data = NULL,
  colour_domain = NULL,
  markers = NULL,
  palette = "YlOrRd",
  line_colour = "grey",
  legend_values = NULL,
  legend_title = NULL,
  add_scale_bar = FALSE,
  reverse_colours = FALSE,
  fill_opacity = 0.75,
  fill_colour = "#ba2d2d",
  fill_colour_weight = 1
)
}
\arguments{
\item{polygon_data}{Polygon data}

\item{colour_domain}{Colour domain for polygon on map}

\item{markers}{A data.frame containing latitude, longitude and label columns}

\item{palette}{Palette, for example YlOrRd or Reds}

\item{legend_values}{Values for legend}

\item{legend_title}{Title for legend}

\item{add_scale_bar}{Add scale bar if TRUE}

\item{fill_colour}{Fill colour of polygon, overriden if colour_domain is passed}

\item{fill_colour_weight}{Polygon colour weight, float from 0 to 1.0, Passed to the weight argument of addPolygons}

\item{polygon_fill_opacity}{Leaflet polygon fill opacity, float from 0 to 1.0, passed to fillOpacity of leaflet::addPolygons}
}
\description{
Create a simple Leaflet map from data
}
