% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mesh.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot a mesh on a Leaflet map}
\usage{
plot_mesh(
  mesh,
  spatial_data = NULL,
  longitude_column = "LONG",
  latitude_column = "LAT",
  markers = NULL
)
}
\arguments{
\item{mesh}{Mesh data}

\item{spatial_data}{Spatial data, either a SpatialPolygonsDataFrame, SpatialPointsDataFrame or an object
that can be converted to a data.frame with longitude and latitude columns}

\item{longitude_column}{Longitude column in spatial_data}

\item{latitude_column}{Latitude column in spatial_data name}

\item{markers}{Markers to display on top of mesh. A named list with latitude, longitude and label names must be given.
Expects longitude name to be longitude, latitude name to be latitude, label name to be label.}
}
\value{
leaflet::leaflet
}
\description{
Plot a mesh on a Leaflet map
}
