% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mesh.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot a mesh}
\usage{
plot_mesh(mesh, point_data, point_colour = "blue", cex = 0.1)
}
\arguments{
\item{mesh}{Mesh data}

\item{point_data}{Points data}

\item{point_colour}{Colour for points}

\item{cex}{Point size magnifier}
}
\value{
terra::plot
}
\description{
Plot a mesh
}
