% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_priors.R
\name{plot_priors_boxplot}
\alias{plot_priors_boxplot}
\title{Create boxplots from priors run data}
\usage{
plot_priors_boxplot(data)
}
\arguments{
\item{data}{}
}
\value{
graphics::boxplot
}
\description{
Create boxplots from priors run data
}
\keyword{internal}
