% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{plot_priors_density}
\alias{plot_priors_density}
\title{Plot density function}
\usage{
plot_priors_density(data, measurement_data)
}
\arguments{
\item{data}{List of parsed model outputs from "parse_model_output" (a list of lists)}

\item{measurement_data}{Measurement data (can be data.frame, SpatialPoints/SpatialPointsDataFrame, or simple feature collection)}
}
\value{
ggplot2::ggplot of measurement density by value
}
\description{
Plot density function
}
