% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot timeseries data}
\usage{
plot_timeseries(
  data,
  x,
  y,
  breaks = NULL,
  x_label = NULL,
  y_label = NULL,
  title = NULL,
  line_colour = "blue",
  horizontal_y = NULL,
  vertical_x = NULL,
  x_lim = NULL,
  y_lim = NULL
)
}
\arguments{
\item{data}{Data to plot (data.frame with x-axis column a character convertible to Date)}

\item{x}{Name of column to plot on x-axis}

\item{y}{Name of column to plot on y-axis}

\item{breaks}{Date break points, must be a sequence of dates}

\item{x_label}{x-axis label}

\item{y_label}{y-axis label}

\item{title}{Figure title}

\item{line_colour}{Line colour}

\item{horizontal_y}{y-intercept for horizontal line}

\item{vertical_x}{x-intercept for vertical line}

\item{x_lim}{Limits for x-axis continuous scale, vector passed to scale_x_continuous}

\item{y_lim}{Limits for y-axis continuous scale, vector passed to scale_y_continuous}

\item{line_width}{Line width}
}
\value{
ggplot
}
\description{
Plot timeseries data
}
