% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_priors.R
\name{priors_shiny}
\alias{priors_shiny}
\title{Interactively set and see the result of different priors}
\usage{
priors_shiny(spatial_data, measurement_data)
}
\arguments{
\item{measurement_data}{}
}
\description{
Interactively set and see the result of different priors
}
\keyword{internal}
