% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_priors.R
\name{priors_shiny}
\alias{priors_shiny}
\title{Interactively set and see the result of different priors}
\usage{
priors_shiny(
  spatial_data,
  measurement_data,
  mesh,
  inla_exposure_param = "Population",
  prior_range = NULL,
  ps_range = NULL,
  prior_sigma = NULL,
  pg_sigma = NULL,
  prior_alpha = NULL,
  pg_alpha = NULL
)
}
\arguments{
\item{spatial_data}{Spatial data}

\item{measurement_data}{Measurement data}

\item{mesh}{INLA mesh}

\item{inla_exposure_param}{Exposure parameter for family = 'poisson' passed to inlabru::bru and then to INLA::inla. Must be column in measurement data.}
}
\value{
shiny::app
}
\description{
Interactively set and see the result of different priors
}
\keyword{internal}
