% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_packages.R
\name{require_packages}
\alias{require_packages}
\title{Checks if a package is installed. If it is it's loaded,
otherwise stop is called with a message telling the user
the package is required.}
\usage{
require_packages(packages)
}
\arguments{
\item{packages}{Package name}
}
\description{
Checks if a package is installed. If it is it's loaded,
otherwise stop is called with a message telling the user
the package is required.
}
\keyword{internal}
