% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_retrieve.R
\name{retrieve_tutorial_data}
\alias{retrieve_tutorial_data}
\title{Retrieve a tutorial dataset and unpacks it to ~/fdmr/tutorial_data}
\usage{
retrieve_tutorial_data(dataset, force_update = FALSE)
}
\arguments{
\item{dataset}{Name of dataset to retrieve}

\item{force_update}{Force retrieval of metadata and dataset}
}
\description{
Retrieve a tutorial dataset and unpacks it to ~/fdmr/tutorial_data
}
