% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_retrieve.R
\name{retrieve_tutorial_data}
\alias{retrieve_tutorial_data}
\title{Retrieve the tutorial datasets and unpack them
to a folder in th user's home directory or a folder
specified in their use config file.}
\usage{
retrieve_tutorial_data(dataset, filename)
}
\arguments{
\item{dataset}{Name of dataset to retrieve, either covid or hydro}
}
\description{
Retrieve the tutorial datasets and unpack them
to a folder in th user's home directory or a folder
specified in their use config file.
}
